refactor(dagster): extract shared dlt probe helpers
Moves the probe/signature/gate primitives out of the powerschool library
into libraries/dlt/probe.py so focus and illuminate can share them.
PowerSchoolTable becomes ProbeTable. No behavior change.

Refs #4447
