+++ src/teamster/code_locations/kippcamden/powerschool/sis/dlt/assets.py
@@ -3,10 +3,8 @@ import pathlib
 import yaml
 
 from teamster.code_locations.kippcamden import CODE_LOCATION
-from teamster.libraries.dlt.powerschool.assets import (
-    PowerSchoolTable,
-    build_powerschool_dlt_assets,
-)
+from teamster.libraries.dlt.powerschool.assets import build_powerschool_dlt_assets
+from teamster.libraries.dlt.probe import ProbeTable
 
 config_file = pathlib.Path(__file__).parent / "config" / "assets.yaml"
 
@@ -14,7 +12,7 @@ assets = [
     build_powerschool_dlt_assets(
         code_location=CODE_LOCATION,
         tables=[
-            PowerSchoolTable(name=a["table_name"], cursor_column=a["cursor_column"])
+            ProbeTable(name=a["table_name"], cursor_column=a["cursor_column"])
             for a in yaml.safe_load(config_file.read_text())["assets"]
         ],
         # dlt's default extract concurrency (5 workers). The single SSH tunnel
