+++ src/teamster/code_locations/kippnewark/powerschool/sis/dlt/sensors.py
@@ -3,10 +3,10 @@ import pathlib
 import yaml
 
 from teamster.code_locations.kippnewark import CODE_LOCATION
-from teamster.libraries.dlt.powerschool.assets import PowerSchoolTable
 from teamster.libraries.dlt.powerschool.sensors import (
     build_powerschool_dlt_intraday_sensor,
 )
+from teamster.libraries.dlt.probe import ProbeTable
 
 config_file = pathlib.Path(__file__).parent / "config" / "assets.yaml"
 
@@ -14,7 +14,7 @@ sensors = [
     build_powerschool_dlt_intraday_sensor(
         code_location=CODE_LOCATION,
         tables=[
-            PowerSchoolTable(name=a["table_name"], cursor_column=a["cursor_column"])
+            ProbeTable(name=a["table_name"], cursor_column=a["cursor_column"])
             for a in yaml.safe_load(config_file.read_text())["assets"]
             if a["intraday"]
         ],
