+++ src/teamster/libraries/dlt/powerschool/assets.py
@@ -1,6 +1,4 @@
 from collections.abc import Iterator
-from dataclasses import dataclass
-from datetime import date, datetime
 
 import dlt
 import sqlalchemy as sa
@@ -14,6 +12,11 @@ from sqlalchemy import Float, Numeric
 from sqlalchemy.types import TypeEngine
 
 from teamster.libraries.dlt.powerschool.resources import OracleResource
+from teamster.libraries.dlt.probe import (
+    ProbeSignatureConfig,
+    ProbeTable,
+    probe_signature,
+)
 from teamster.libraries.ssh.resources import SSHResource
 
 # PowerSchool tables are owned by the `ps` schema. The ODBC pipeline reaches
@@ -37,87 +40,12 @@ def _asset_key(code_location: str, table_name: str) -> AssetKey:
     return AssetKey([code_location, _SOURCE_NAME, "sis", table_name])
 
 
-@dataclass(frozen=True)
-class PowerSchoolTable:
-    """One PowerSchool table's sync config.
-
-    cursor_column None means the table has no change-tracking column and is
-    always fully replaced when selected (nightly tier only).
-    """
-
-    name: str
-    cursor_column: str | None
-
-
-def probe_signature(
-    connection, table_name: str, cursor_column: str | None
-) -> dict[str, int | str | None]:
-    """Fetch the change signature for a table: total count + max cursor.
-
-    Equality-compared against the stored signature; drift in either value
-    (including a cursor regression) triggers a full replace. Tables without a
-    cursor column are count-only — the signature still carries
-    ``max_cursor: None`` so it compares equal to the run-config round-trip
-    shape (which defaults the key to None). Values are JSON-serializable for
-    dlt resource state.
-    """
-    if cursor_column is None:
-        (count,) = connection.execute(
-            # trunk-ignore(bandit/B608): table name from static YAML config
-            sa.text(f"SELECT COUNT(*) FROM {table_name}")
-        ).one()
-
-        return {"count": int(count), "max_cursor": None}
-
-    count, max_cursor = connection.execute(
-        # trunk-ignore(bandit/B608): table/column names from static YAML config
-        sa.text(f"SELECT COUNT(*), MAX({cursor_column}) FROM {table_name}")
-    ).one()
-
-    if max_cursor is None:
-        max_cursor_value = None
-    elif isinstance(max_cursor, (datetime, date)):
-        max_cursor_value = max_cursor.isoformat()
-    else:
-        # Non-temporal cursor (e.g. a numeric change column on a future table);
-        # store its string form so the signature stays JSON-serializable.
-        max_cursor_value = str(max_cursor)
-
-    # int(count): mirror the JSON-safe-scalar normalization done for max_cursor
-    # above (oracledb returns int today, but keep the state doc driver-agnostic).
-    return {"count": int(count), "max_cursor": max_cursor_value}
-
-
-def _compute_changed(
-    selected: list[PowerSchoolTable],
-    current: dict[str, dict],
-    stored: dict[str, dict],
-) -> list[PowerSchoolTable]:
-    """Select tables whose just-probed signature differs from the stored one.
-
-    Drift in count or max cursor — or a missing stored entry (first tick, or a
-    table new to intraday) — selects the table. No-cursor tables carry a
-    count-only signature (``max_cursor: None``), so a net row add/remove
-    selects them; in-place edits are caught by the nightly full refresh.
-    """
-    return [
-        table for table in selected if current.get(table.name) != stored.get(table.name)
-    ]
-
-
 def _resolve_extract_workers(tag_value: str | None, param: int | None) -> int | None:
     """Resolve the dlt extract worker cap: a per-run tag overrides the
     factory param, which overrides dlt's default (None = leave default)."""
     if tag_value is not None:
         return int(tag_value)
     return param
-
-
-class ProbeSignatureConfig(Config):
-    """One table's probed change signature, passed by the intraday sensor."""
-
-    count: int
-    max_cursor: str | None = None
 
 
 class PowerSchoolDltConfig(Config):
@@ -181,7 +109,7 @@ class PowerSchoolDagsterDltTranslator(DagsterDltTranslator):
 
 
 def _build_resource(
-    table: PowerSchoolTable, signature: dict | None, connection_url: str, arraysize: int
+    table: ProbeTable, signature: dict | None, connection_url: str, arraysize: int
 ):
     """Build one full-replace, parallel-extracted dlt resource for a table.
 
@@ -237,7 +165,7 @@ def _build_resource(
 
 @dlt.source(name=_SOURCE_NAME)
 def _powerschool_source(
-    selected: list[PowerSchoolTable],
+    selected: list[ProbeTable],
     signatures: dict[str, dict],
     connection_url: str,
     arraysize: int,
@@ -251,7 +179,7 @@ def _powerschool_source(
 
 def build_powerschool_dlt_assets(
     code_location: str,
-    tables: list[PowerSchoolTable],
+    tables: list[ProbeTable],
     op_tags: dict[str, object] | None = None,
     max_extract_workers: int | None = None,
 ):
@@ -381,15 +309,3 @@ def build_powerschool_dlt_assets(
                 raise
 
     return _assets
-
-
-def _stored_signatures(dlt_pipeline, source_name: str) -> dict[str, dict]:
-    """Read last-run per-resource signatures from dlt pipeline state."""
-    resources = (
-        dlt_pipeline.state.get("sources", {}).get(source_name, {}).get("resources", {})
-    )
-    return {
-        name: res_state["signature"]
-        for name, res_state in resources.items()
-        if isinstance(res_state, dict) and "signature" in res_state
-    }
